Fragmented parts of the website to components & Added Blogs
